Made changes to the code
