+++ src/components/StoriesCard/FeaturedStories.tsx
@@ -1,15 +1,41 @@
+import { useNavigate } from "react-router-dom";
 import photo from "../../../public/images/huldu1 1.png";
 
 function getRandomPhoto() {
   return photo;
 }
 
-export const FeaturedStories = ({ item }) => {
+export const FeaturedStories = ({ item, categoryName }) => {
+
+  let navigate = useNavigate();
+  console.log(item);
+
+  const handleStoryClick = (e: string, categoryName: string) => {
+    if (categoryName === "Allt") {
+      categoryName = "all"
+    }
+    if (categoryName === "Tröll") {
+      categoryName = "troll"
+    }
+
+    if (categoryName === "Draugar") {
+      categoryName = "draugar"
+    }
+
+    if (categoryName === "alfar-og-huldufolk") {
+      categoryName = "alfa";
+    }
+
+    if (categoryName === "Helgisögur") {
+      categoryName = "ur-efra-og-nedra-helgisogur"
+    }
+    navigate(`/stories/${categoryName}/${e}`);
+  }
+
   return (
     <div className="relative text-center">
       <img src={getRandomPhoto()} alt={item} />
-
-      <h2>{item}</h2>
+      <h2 onClick={(e) => handleStoryClick(e.target.innerText, categoryName)}>{item == "categories" ? null : item.replace(/[/]/g, "")}</h2>
     </div>
   );
 };
